crawl-tool/requirements.txt
appdirs==1.4.4
asgiref==3.5.0
attrs==21.4.0
Automat==20.2.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
dateparser==1.1.0
demjson3==3.0.5
dj-database-url==0.5.0
Django==4.0.2
django-serverside-datatable==2.0.0
facebook-scraper==0.2.56
fake-useragent==0.1.11
gunicorn==20.1.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.2
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.8.0
parse==1.19.0
parsel==1.6.0
priority==1.3.0
Protego==0.2.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyDispatcher==2.0.5
pyee==8.2.2
pymongo==4.0.1
pyOpenSSL==22.0.0
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
queuelib==1.6.2
regex==2022.1.18
requests==2.27.1
requests-html==0.10.0
Scrapy==2.5.1
scrapy-splash==0.8.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
tqdm==4.63.0
Twisted==22.1.0
typing-extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
Unipath==1.1
urllib3==1.26.8
w3lib==1.22.0
websockets==10.2
whitenoise==6.0.0
zipp==3.7.0
zope.interface==5.4.0
webdriver-manager
selenium-wire
